Add custom validator for highlight for JSON
By default, models.JSONField provided by django doesn't validate input
based on any schema. Was the source of a pretty annoying TypeError that
I was dealing with earlier.

Using jsonschema library, implemented a custom JSONSchemaValidator that
validates JSON input according to a pre-defined schema.

The main source used was https://stackoverflow.com/a/49036841/, but my
implementation needed to be adapted for DRF.