Refactor highlight field from JSONField to ints
Decided to use individual columns for highlight field instead of
JSONField.

Inspired by GPT-4 when I asked it to serialize a JSON object. Expected
it to come up with a custom validator, but it simply used serializers.

Realized that if my JSON schema isn't changing, it doesn't make sense to
use JSONField. The main advantage is the flexibility of accepting
different JSON objects.

Additionally, the upside of using individual columns is that:
1) Improved database performance
2) Not dependent on Postgres having JSON type
3) Will make editing annotations easier (if underlying article changes)

The downside of moving away from JSONField is that
1) Application needs to handle flattening and unflattening

Instead of making django handle the flattening / unflattening, decided
to let client worry about it.