Document IsOwnerOrReadOnly behvaiour
Noting that it does not get automatically called on generic views that
return lists (e.g. ones that use .get_queryset)

Rather, it will be called in views that use .get_object. In the event
that we want to customize a .get_object method by overriding it, we must
explicitly call .check_object_permissions to run a method like
.has_object_permissions in IsOwnerOrReadOnly.