feat: add IsAuthenticated permish Bookmark view
Was letting user create a bookmark.

The IsOwner custom permission works for check_object_permissions(), but
it doesn't work for List or Create requests (which is now a part of this
view).

Add the `IsAuthenticated` permission which works for List or Create
requests. Note now that `IsAuthenticated` and `IsOwner` do not actually
override eachtoer. They are separate permissions. The former for the
request, the latter for the object.